requirements.txt
Django==4.2
djangorestframework==3.14.0
kubernetes==26.1.0
minio==7.1.15
psycopg2_binary==2.9.6
PyYAML==6.0
PyYAML==6.0
